simple history made, need to implement the clickability

+++ includes/token/function.h
@@ -25,7 +25,7 @@ public:
 
     // Evaluates with already resolved double arguments from RPN
     double evaluate(const vector<double>& args) const {
-        if (args.size() != static_cast<size_t>(_args_count)) {
+        if (args.size() != _args_count) {
             throw runtime_error("Function '" + _func_name + "' called with " + std::to_string(args.size()) + " arguments, but expects " + std::to_string(_args_count));
         }
 
@@ -38,7 +38,7 @@ public:
             double arg1 = args[0];
             double arg2 = args[1]; 
             if (_func_name == "pow") {
-                return std::pow(arg1, arg2); 
+                return pow(arg1, arg2); 
             }
             
         }
